test: Moved ReplicationFailoverIntegration
