+++ build.gradle.kts
@@ -387,6 +387,7 @@ tasks.register<Test>("in-container-aurora") {
     filter.includeTestsMatching("testsuite.integration.container.AuroraMysqlIntegrationTest")
     filter.includeTestsMatching("testsuite.integration.container.AuroraMysqlFailoverIntegrationTest")
     filter.includeTestsMatching("testsuite.integration.container.AuroraMysqlHardFailureIntegrationTest")
+    filter.includeTestsMatching("testsuite.integration.container.ReplicationFailoverIntegrationTest")
 }
 
 tasks.register<Test>("in-container-aurora-performance") {
